requirements.txt
Django==1.8.5
Pillow==2.9.0
argparse==1.2.1
django-crispy-forms==1.5.2
django-extensions==1.5.7
django-grappelli==2.7.1
django-pgjson==0.3.1
ipython==1.2.1
psycopg2==2.6.1
six==1.10.0
wsgiref==0.1.2